+++ includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/02 14:57:03 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/08 14:04:38 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/19 14:18:20 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,6 @@
 
 void	free_string_double_array(char **str);
 char	*add_slash(char *path);
+void	print_msg(char *str);
 
 #endif
