+++ srcs/create_token_util.c
@@ -6,66 +6,11 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/13 15:10:32 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/18 15:52:32 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/19 14:27:31 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"../includes/minishell.h"
-
-// bool	create_token_word(int *start, char *line, t_token **head)
-// {
-// 	char	*word;
-// 	int		j;
-// 	int		len;
-// 	bool	err_flg;
-
-// 	err_flg = false;
-// 	len = (int)ft_strlen(line);
-// 	j = *start + 1;
-// 	while (!is_split_char(line[j]) && !is_quart(line[j]) && j < len)
-// 		j++;
-// 	word = ft_substr(line, *start, j - *start);
-// 	if (!word)
-// 		return (true);
-// 	if (*start == 0 || is_split_char(line[*start - 1]))
-// 		err_flg = add_token(head, create_token_node(word));
-// 	else
-// 		err_flg = add_current_token(word, head);
-// 	*start = j;
-// 	return (err_flg);
-// }
-
-t_token	*create_token_node(char *word)
-{
-	t_token	*token;
-
-	token = (t_token *)ft_calloc(sizeof(t_token), 1);
-	if (!token)
-		return (NULL);
-	token->word = word;
-	token->next = NULL;
-	return (token);
-}
-
-bool	add_token(t_token **head, t_token *new)
-{
-	t_token	*temp;
-	bool	err_flg;
-
-	err_flg = false;
-	if (new == NULL || head == NULL)
-		return (true);
-	if (*head == NULL)
-		*head = new;
-	else
-	{
-		temp = *head;
-		while (temp->next)
-			temp = temp->next;
-		temp->next = new;
-	}
-	return (err_flg);
-}
 
 bool	is_split_char(char c)
 {
@@ -74,7 +19,7 @@ bool	is_split_char(char c)
 	return (false);
 }
 
-bool	is_quart(char c)
+bool	is_quote(char c)
 {
 	if (c == '\'' || c == '\"')
 		return (true);
@@ -94,4 +39,14 @@ bool	add_current_token(char *word, t_token **head)
 		return (true);
 	get_last_token(head)->word = temp;
 	return (err_flg);
+}
+
+t_token	*get_last_token(t_token **head)
+{
+	t_token	*temp;
+
+	temp = *head;
+	while (temp->next)
+		temp = temp->next;
+	return (temp);
 }
