+++ srcs/utils/print_msg.c
@@ -6,11 +6,11 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/16 15:41:25 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/16 15:43:35 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/19 14:17:56 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 
 void	print_msg(char *str)
 {
